requirements.txt
flask==2.0.1
flask-cors==3.0.10
pytesseract==0.3.8
Pillow==8.3.1
werkzeug==2.0.1
click==8.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 